bookworm_api/requirements.txt
fastapi>=0.68.0
pydantic>=1.8.0
openai>=1.0.0
torch>=2.0.0
transformers>=4.30.0
pinecone-client>=2.2.0
python-dotenv>=0.19.0
uvicorn>=0.15.0